# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/musicos": {
      "post": {
        "operationId": "MusicoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMusicianDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Musician created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - host only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new musician (host only)",
        "tags": [
          "Musicians"
        ]
      },
      "get": {
        "operationId": "MusicoController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of musicians"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all musicians",
        "tags": [
          "Musicians"
        ]
      }
    },
    "/musicos/{id}": {
      "patch": {
        "operationId": "MusicoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMusicianDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Musician updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - can only update own profile unless host"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update musician (self or host)",
        "tags": [
          "Musicians"
        ]
      },
      "delete": {
        "operationId": "MusicoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Musician deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - host only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete musician (host only)",
        "tags": [
          "Musicians"
        ]
      }
    },
...
    "/spotify/import": {
      "post": {
        "description": "If 'jamId' is provided, tracks are appended to the existing jam. If omitted, a new jam is created. When importing to an existing jam, tracks already present are skipped.",
        "operationId": "SpotifyController_importPlaylist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Import completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid playlist URL"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - user is not the jam host"
          },
          "404": {
            "description": "Jam or playlist not found"
          },
          "503": {
            "description": "Spotify integration not configured"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import a Spotify playlist as a new jam or append to existing jam",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/spotify/export": {
      "post": {
        "operationId": "SpotifyController_exportPlaylist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Spotify playlist created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResultDto"
                }
              }
            }
          },
          "400": {
            "description": "No valid Spotify links in jam"
          },
          "401": {
            "description": "Unauthorized or invalid Spotify token"
          },
          "404": {
            "description": "Jam not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export a jam as a Spotify playlist",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/spotify/track": {
      "post": {
        "operationId": "SpotifyController_getTrackMetadata",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTrackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Track metadata retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackMetadataDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid track URL or URI"
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Spotify integration not configured"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Spotify track metadata from URL or URI",
        "tags": [
          "Spotify"
        ]
      }
    }
  },
  "info": {
    "title": "Karaoke Jam API",
    "description": "Backend API para organização de Jam Sessions presenciais em tempo real",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MusicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "neededDrums": {
            "type": "number",
            "example": 0,
            "description": "Number of drummers needed"
          },
          "neededGuitars": {
            "type": "number",
            "example": 0,
            "description": "Number of guitarists needed"
          },
          "neededVocals": {
            "type": "number",
            "example": 0,
            "description": "Number of vocalists needed"
          },
          "neededBass": {
            "type": "number",
            "example": 0,
            "description": "Number of bassists needed"
          },
          "neededKeys": {
            "type": "number",
            "example": 0,
            "description": "Number of keyboardists needed"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "status",
          "neededDrums",
          "neededGuitars",
          "neededVocals",
          "neededBass",
          "neededKeys",
          "createdAt"
        ]
      },
      "MusicianResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "instrument": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "supabaseUserId": {
            "type": "string"
          },
          "isHost": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isHost",
          "createdAt"
        ]
      },
      "RegistrationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "musicianId": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "jamMusicId": {
            "type": "string"
          },
          "scheduleId": {
            "type": "string"
          },
          "instrument": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "musician": {
            "$ref": "#/components/schemas/MusicianResponseDto"
          }
        },
        "required": [
          "id",
          "musicianId",
          "jamId",
          "jamMusicId",
          "status",
          "createdAt"
        ]
      },
      "JamMusicaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "musicId": {
            "type": "string"
          },
          "music": {
            "$ref": "#/components/schemas/MusicaResponseDto"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          }
        },
        "required": [
          "id",
          "jamId",
          "musicId",
          "music"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jamId": {
            "type": "string"
          },
          "musicId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUGGESTED",
              "SCHEDULED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "pausedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "music": {
            "$ref": "#/components/schemas/MusicaResponseDto"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          }
        },
        "required": [
          "id",
          "jamId",
          "musicId",
          "order",
          "status",
          "createdAt"
        ]
      },
      "JamResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "hostName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "qrCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LIVE",
              "FINISHED"
            ]
          },
          "playbackState": {
            "type": "string",
            "enum": [
              "STOPPED",
              "PLAYING",
              "PAUSED"
            ]
          },
          "currentScheduleId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "jamMusics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JamMusicaResponseDto"
            }
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationResponseDto"
            }
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateMusicianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Musician name"
          },
          "contact": {
            "type": "string",
            "description": "Musician contact"
          },
          "instrument": {
            "type": "string",
            "description": "Primary instrument"
          },
          "level": {
            "type": "string",
            "description": "Experience level",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "PROFESSIONAL"
            ]
          }
        },
        "required": [
          "name",
          "contact",
          "instrument",
          "level"
        ]
      },
      "UpdateMusicianDto": {
        "type": "object",
        "properties": {}
      },
...
      "ImportPlaylistDto": {
        "type": "object",
        "properties": {
          "playlistUrl": {
            "type": "string",
            "description": "Spotify playlist URL or URI"
          },
          "jamId": {
            "type": "string",
            "description": "Jam ID to import into (if omitted, creates new jam)"
          },
          "name": {
            "type": "string",
            "description": "Jam name (defaults to playlist name, ignored if jamId provided)"
          },
          "description": {
            "type": "string",
            "description": "Jam description (defaults to playlist description, ignored if jamId provided)"
          },
          "date": {
            "type": "string",
            "description": "Jam date in ISO 8601 format (ignored if jamId provided)"
          },
          "location": {
            "type": "string",
            "description": "Jam location (ignored if jamId provided)"
          }
        },
        "required": [
          "playlistUrl"
        ]
      },
      "ImportResultDto": {
        "type": "object",
        "properties": {
          "jam": {
            "description": "Target jam (existing or newly created)",
            "allOf": [
              {
                "$ref": "#/components/schemas/JamResponseDto"
              }
            ]
          },
          "importedTracks": {
            "type": "number",
            "description": "Number of new Music records created"
          },
          "reusedTracks": {
            "type": "number",
            "description": "Number of existing Music records reused"
          },
          "skippedTracks": {
            "type": "number",
            "description": "Number of tracks that failed to import"
          },
          "addedTracks": {
            "type": "number",
            "description": "Number of tracks added to the jam"
          },
          "duplicateTracks": {
            "type": "number",
            "description": "Number of tracks skipped (already in jam)"
          },
          "isExistingJam": {
            "type": "boolean",
            "description": "Whether an existing jam was used"
          },
          "errors": {
            "description": "Error details for skipped tracks",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "jam",
          "importedTracks",
          "reusedTracks",
          "skippedTracks",
          "addedTracks",
          "duplicateTracks",
          "isExistingJam"
        ]
      },
      "ExportPlaylistDto": {
        "type": "object",
        "properties": {
          "jamId": {
            "type": "string",
            "description": "Jam ID to export"
          },
          "spotifyAccessToken": {
            "type": "string",
            "description": "Spotify OAuth access token from frontend"
          },
          "playlistName": {
            "type": "string",
            "description": "Playlist name (defaults to jam name)"
          },
          "playlistDescription": {
            "type": "string",
            "description": "Playlist description"
          },
          "public": {
            "type": "boolean",
            "description": "Whether the playlist should be public",
            "default": false
          }
        },
        "required": [
          "jamId",
          "spotifyAccessToken"
        ]
      },
      "ExportResultDto": {
        "type": "object",
        "properties": {
          "spotifyPlaylistId": {
            "type": "string",
            "description": "Spotify playlist ID"
          },
          "spotifyPlaylistUrl": {
            "type": "string",
            "description": "Spotify playlist URL"
          },
          "totalTracks": {
            "type": "number",
            "description": "Total tracks added to playlist"
          },
          "skippedTracks": {
            "type": "number",
            "description": "Tracks skipped (no valid Spotify link)"
          },
          "errors": {
            "description": "Error details for skipped tracks",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "spotifyPlaylistId",
          "spotifyPlaylistUrl",
          "totalTracks",
          "skippedTracks"
        ]
      },
      "GetTrackDto": {
        "type": "object",
        "properties": {
          "trackUrl": {
            "type": "string",
            "description": "Spotify track URL or URI",
            "example": "https://open.spotify.com/track/4iV5W9uYEdYUVa79Axb7Rh"
          }
        },
        "required": [
          "trackUrl"
        ]
      },
      "TrackMetadataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Spotify track ID"
          },
          "title": {
            "type": "string",
            "description": "Track title"
          },
          "artist": {
            "type": "string",
            "description": "Artist name(s), comma-separated if multiple"
          },
          "durationMs": {
            "type": "number",
            "description": "Track duration in milliseconds"
          },
          "spotifyUrl": {
            "type": "string",
            "description": "Spotify URL for the track"
          },
          "albumName": {
            "type": "string",
            "description": "Album name"
          },
          "albumImageUrl": {
            "type": "string",
            "description": "Album cover image URL"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "durationMs",
          "spotifyUrl"
        ]
      }
    }
  }
}